Move ActionCopyToClipboard component to malwarefront ui
