+++ malwarecage/web/src/commons/ui/ActionCopyToClipboard.js
