+++ malwarecage/web/src/commons/ui/ui.js
@@ -16,6 +16,7 @@ export { default as ProtectedRoute } from "./ProtectedRoute";
 export { default as RefString } from "./RefString";
 export { default as SortedList } from "./SortedList";
 export { default as View } from "./View";
+export { default as ActionCopyToClipboard } from "./ActionCopyToClipboard";
