+++ malwarecage/web/src/components/Attributes.js
@@ -4,8 +4,7 @@ import AttributesAddModal from './AttributesAddModal';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import api from "@malwarefront/api";
 import { fromPlugin, Extendable } from "@malwarefront/extensions";
-import { DataTable, ConfirmationModal } from "@malwarefront/ui";
-import ActionCopyToClipboard from "./ActionCopyToClipboard";
+import { DataTable, ConfirmationModal, ActionCopyToClipboard } from "@malwarefront/ui";
 
 let attributeRenderers = {}
