+++ malwarecage/web/src/components/RecentView/RecentInnerRow.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ActionCopyToClipboard from "../ActionCopyToClipboard";
+import { ActionCopyToClipboard } from "@malwarefront/ui";
 
 export default function RecentInnerRow(props) {
     const classNames = (
