+++ malwarecage/web/src/components/Relations.js
@@ -3,9 +3,8 @@ import {Link} from 'react-router-dom';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {TagList} from './Tag';
 import api from "@malwarefront/api";
-import { ObjectLink } from '@malwarefront/ui';
+import { ObjectLink, ActionCopyToClipboard } from '@malwarefront/ui';
 import RelationsAddModal from "./RelationsAddModal";
-import ActionCopyToClipboard from "./ActionCopyToClipboard";
 
 class RelationsBox extends Component {
     state = {
