+++ malwarecage/web/src/components/ShowConfig.js
@@ -8,9 +8,8 @@ import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import api from "@malwarefront/api";
 import { makeSearchConfigElementLink, makeSearchLink, makeSearchDateLink, downloadData } from "@malwarefront/helpers";
 import { Extension } from "@malwarefront/extensions";
-import { DataTable, DateString, ObjectLink, View, HexView } from "@malwarefront/ui";
+import { DataTable, DateString, ObjectLink, View, HexView, ActionCopyToClipboard } from "@malwarefront/ui";
 import ShowObjectPresenter from './ShowObjectPresenter';
-import ActionCopyToClipboard from "./ActionCopyToClipboard";
 
 class ConfigRow extends Component {
     constructor(props){
