+++ malwarecage/web/src/components/ShowSample.js
@@ -7,10 +7,9 @@ import {ConfigTable} from "./ShowConfig";
 import api from "@malwarefront/api";
 import {makeSearchLink, makeSearchDateLink} from "@malwarefront/helpers";
 import { Extendable } from "@malwarefront/extensions";
-import { Identicon, DataTable, DateString, Hash, View } from "@malwarefront/ui";
+import { Identicon, DataTable, DateString, Hash, View, ActionCopyToClipboard } from "@malwarefront/ui";
 import ShowObjectPresenter, {joinActions} from './ShowObjectPresenter';
 import ShowSamplePreview from "./ShowSamplePreview";
-import ActionCopyToClipboard from "./ActionCopyToClipboard";
 
 function SampleDetails(props) {
     return (
